finish states to seperate store
